requirements.txt
# Core

openai>=1.0.0

python-dotenv>=1.0.0



# PDF Processing

pypdf2>=3.0.0

pdfplumber>=0.10.0



# HTML Processing

beautifulsoup4>=4.12.0

html2text>=2024.2.26

lxml>=5.0.0



# Spreadsheets

openpyxl>=3.1.0



# Images

pillow>=10.0.0



# Text Processing

tiktoken>=0.5.0  # Para contagem de tokens e chunking



# Utilities

requests>=2.31.0  # Para download de arquivos via URL
